// src/common/types.tsx
export enum Color {
  BLUE = "blue",
  BLUE_LIGHT = "blue_light",
  ORANGE = "orange",
  ORANGE_LIGHT = "orange_light",
  YELLOW = "yellow",
  YELLOW_LIGHT = "yellow_light",
  GREEN = "green",
  GREEN_LIGHT = "green_light",
  PURPLE = "purple",
  PURPLE_LIGHT = "purple_light",
  BROWN = "brown",
  GRAY = "gray",
  BLACK = "black",
}